Relocated global components
